requirements.txt
altair==4.0
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
boto3==1.18.41
botocore==1.21.41
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
exceptiongroup==1.1.2
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.21.5
outcome==1.2.0
packaging==23.1
pandas==1.4.2
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plotly==5.11.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
s3transfer==0.5.2
selenium==4.8.0
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
streamlit==1.15.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
wsproto==1.2.0
zipp==3.16.2
